Improve Deployment creations screen - #5734 (#6214)
* Improve workload deployments view.
