+++ shell/assets/translations/en-us.yaml
@@ -5083,7 +5083,7 @@ workload:
       livenessProbe: Liveness Check
       livenessTip: Containers will be restarted when this check is failing.  Not recommended for most uses.
       noHealthCheck: "There is not a Readiness Check, Liveness Check or Startup Check configured."
-      readinessProbe: Readiness Check
+      readinessProbe: Readiness Checks
       readinessTip: Containers will be removed from service endpoints when this check is failing.  Recommended.
       startupProbe: Startup Check
       startupTip: Containers will wait until this check succeeds before attempting other health checks.
@@ -5158,6 +5158,7 @@ workload:
       protocol: Protocol
       listeningPort: Listening Port
     removeContainer: Remove Container
+    addContainer: Add Container
     security:
       addCapabilities: Add Capabilities
       addGroupIDs: Add Group IDs
@@ -5194,6 +5195,7 @@ workload:
     standard: Standard Container
     terminationState: "Last state: Terminated with {lsExitCode}: {lsDescription}, started: {lsStartedAt}, finished: {lsFinishedAt}"
     titles:
+      pods: Pods
       container: Container
       command: Command
       containers: Containers
