+++ shell/components/Tabbed/Tab.vue
@@ -89,10 +89,27 @@ export default {
     :aria-hidden="!active"
     role="tabpanel"
   >
-    <h2 v-if="shouldShowHeader">
-      {{ label }}
-      <i v-if="tooltip" v-tooltip="tooltip" class="icon icon-info icon-lg" />
-    </h2>
+    <div v-if="shouldShowHeader" class="tab-header">
+      <h2>
+        {{ label }}
+        <i v-if="tooltip" v-tooltip="tooltip" class="icon icon-info icon-lg" />
+      </h2>
+      <slot name="tab-header-right" />
+    </div>
     <slot v-bind="{active}" />
   </section>
 </template>
+
+<style lang="scss" scoped>
+.tab-header {
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 15px;
+  align-items: center;
+
+  h2 {
+    margin: 0;
+
+  }
+}
+</style>
